+++ CHANGELOG.md
@@ -20,6 +20,20 @@ For more information on SemVer, please visit http://semver.org/.
 
 ---
 
+## Version 1.7.1
+### 2014-11-05
+
+#### Bugs
+
+* Fixed wrong enqueue components for preferences
+
+#### Enhancements
+
+* Improved `WPDKUIControlButton` class with toggle feature (see https://github.com/wpXtreme/wpdk/wiki/Button-Toggle-UI-Control)
+* Added `wpdkProgressBars()` methods to manage the progress bar (see https://github.com/wpXtreme/wpdk/wiki/WPDKUIProgress)
+* Added `WPDKUIProgress`
+* Update WPDK UI Progress styles with Bootstrap v3.3.0
+
 ## Version 1.7.0
 ### 2014-10-29